refactor: split handler.go into domain files, add Client interface and middleware
Phase 2 structural refactoring:
- Split 1500-line handler.go into 8 domain files (metrics, fields, alerts,
  dashboards, services, query_builder, logs, traces)
- Extract inline closures into named handler methods (handleListMetrics, etc.)
- Define Client interface in internal/client/interface.go for testability
- Change GetClient return type to Client interface
- Add tool handler middleware for centralized logging with tool name and duration
